Move _sanitizeCreateData() to custom transaction entity on chain module
